Improving build script

+++ build.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 #
+# Remove any previous archives
+rm -f lamda_authorizer-*.zip
+
+# Run tests
+npm test
+if [ ! $? == 0 ]; then
+    exit 1
+fi
+
 # Clean and npm install for current deps
 rm -rf node_modules
-npm install
+npm install --production
 
 # Build version number
 VERSION="$1"
@@ -15,3 +24,6 @@ fi
 # Create archive with node.js module
 ARCHIVE_NAME="lamda_authorizer-${RELEASE_VERSION}.zip"
 zip $ARCHIVE_NAME *.js lib/*.js node_modules
+
+# Reinstall all the deps for development to continue
+npm install
